requirements-new.txt
alabaster==0.7.12
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.12.12
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.6
coverage==6.5.0
crispy-bootstrap5==0.7
cryptography==38.0.3
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
Django==4.0.8
django-ajax-datatable==4.4.4
django-filter==22.1
django-allauth==0.51.0
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-betterforms==2.0.0
django-celery-beat==2.4.0
django-pure-pagination==0.3.0
django-coverage-plugin==2.0.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-dynamic-forms==0.5.3
django-environ==0.9.0
django-extensions==3.2.1
django-formtools==2.4
django-htmx==1.13.0
django-jsignature==0.11
django-model-utils==4.2.0
django-redis==5.2.0
django-smart-selects==1.6.0
django-stubs==1.12.0
django-stubs-ext==0.7.0
django-tables2==2.4.1
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
docutils==0.19
et-xmlfile==1.1.0
exceptiongroup==1.0.1
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.1
filelock==3.8.0
flake8==5.0.4
flake8-isort==5.0.0
fonttools==4.38.0
hiredis==2.0.0
identify==2.5.8
idna==3.4
imagesize==1.4.1
iniconfig==1.1.1
install==1.3.5
ipdb==0.13.9
ipython==8.6.0
isort==5.10.1
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
lazy-object-proxy==1.8.0
livereload==2.6.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.5
oauthlib==3.2.2
openpyxl==2.4.11
packaging==21.3
pandas==1.5.1
pandasql==0.7.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.5
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pyquery==1.4.3
pytest==7.2.0
pytest-django==4.5.2
pytest-sugar==0.9.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.6.0
termcolor==2.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.5.0
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.6
watchdog==2.1.9
watchfiles==0.18.1
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
django-group-by==0.3.1